Fix put_pixel coordinates for gdk-pixbuf on Arch.
Pixbuf::put_pixel takes u32; the colour wash was still looping as i32,
so 0.23.2 failed to compile on the user's machine.

+++ crates/vinilo/src/components/artwork.rs
@@ -221,8 +221,8 @@ fn tone_wash(color: [u8; 3], size: i32) -> Option<gdk_pixbuf::Pixbuf> {
     let color = lift_chroma(color);
     let n = size as f32;
     let cx = (n - 1.0) / 2.0;
-    for y in 0..size {
-        for x in 0..size {
+    for y in 0..size as u32 {
+        for x in 0..size as u32 {
             let dx = (x as f32 - cx) / n;
             let dy = (y as f32 - cx) / n;
             let t = (dx * dx + dy * dy).sqrt() * 1.35;
